requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
httpx==0.24.1
pydantic==2.4.2
langchain-openai==0.0.2
azure-identity==1.15.0
azure-ai-documentintelligence==1.0.0b1
Werkzeug==2.3.7